﻿[DotNetRCAndUtils:702] Added a new file supplier should be added which allows files to be supplied via the use of a context menu in the windows shell.
When such a supplier is running, by right-clicking on appropriate files/folders in the Windows shell (file explorer), a "Send to FAM" sub-menu should become available. There should be one item for every running context menu file supplier on the machine.

The supplier's configuration should allow for limits on the cases when the menu option is available:
- Allow only files that match a specified filter
- Allow or disallow matching files from within selected folders to be included (recursively).
- Allow only files where the start of the full file path matches a specified value (To ensure files be queued via a UNC patch, for example).

If a selection in Windows explorer does not contain any items that meet these requirements, the context menu option will not appear. If only some of the files qualify, the menu option will appear, but it will only supply the qualified files when selected.

In order to allow this to work, there will be a new service ESIPCService which will need to be running for the context menus to appear. Ideally this is something we want to make happen transparently to the user; some discussion is needed on what is the best way to accomplish that.

Reviewed by:
William (6/4)

[git-vault-id] Extract$/Engineering@15933/17032