﻿[LegacyRCAndUtils:6046] In order to prevent UAC issues, adjusted TestHarness and IDShieldTester so that they output results and statistics to the application data directory rather than beneath “C:\Program Files\Extract Systems\TestingFiles” and fixed a couple automated tests that were needlessly generating VOA files.
Note that IDShieldTester statistics will continue to be output to directories that may present UAC issues if specifically directed to do so via the “OutputFilesFolder” dat file parameter or the IDShieldStatisticsReporter.

Also, turned on warnings as errors for the Extract.Utilities.ShellExtensions project and fixed a couple situations that were generating errors.

Reviewed by:
William (0/0)

[git-vault-id] Extract$/Engineering@15951/17050