﻿[LegacyRCAndUtils:6043] FDRS now expects the DetectAndReportFailure.ini file to be at [CommonAppData]\Extract Systems\FDRS. Note that this SCR is being left open for the time being since there is still a corresponding installer change that needs to occur.
[LegacyRCAndUtils:6162]
The FAM title bar should be changed to:
"<FPSFileName> - <DatabaseName> on <DBServerName> - File Action Manager"

[LegacyRCAndUtils:6163]
In order to prevent the database from being cleared while there are FAMs that are queuing or displaying statistics, the DB schema has been changed so that the ProcessingFAM table is now the ActiveFAM table. The table now has two additional columns: “Queuing” and “Processing” to indicate whether the FAMs that are currently running are queuing or processing (if both columns are false, the FAM is only reporting statistics).

Reviewed by:
William (0/0)

[git-vault-id] Extract$/Engineering@16002/17104