﻿[LegacyRCAndUtils:6156, 6158] The check state of “Show application traces” and “Show displayed exceptions” is no longer persisted between UEX sessions.
[LegacyRCAndUtils:6172]
The 5 minute lower limit of AutoRevertTimeOutInMinutes is no longer enforced in debug mode.

[FlexIDSCore:2587]
The document classifier now has an option to indicate that when document classification results already exist, whether to skip running it again and use the existing classification or whether to clear the existing results and re-run the classifier.

This option will default to checked in existing rulesets to preserve existing behavior, but will default to unchecked in new rulesets.

Reviewed by:
William (0/1)

[git-vault-id] Extract$/Engineering@16171/17315