﻿$/Engineering/ProductDevelopment/AttributeFinder/Build $/Engineering/ProductDevelopment/Common
$/Engineering/ProductDevelopment/Common/InstallScripts
$/Engineering/ProductDevelopment/AttributeFinder/Installation/UCLID FlexIndex
$/Engineering/ProductDevelopment/AttributeFinder/Installation/UCLID FlexIndex SDK
$/Engineering/ProductDevelopment/AttributeFinder/Installation/FKBInstall
$/Engineering/ProductDevelopment/AttributeFinder/IndustrySpecific/Redaction/Installation/IDShield

FlexIDSCore #4810
Modified installation to prompt for uninstall - this was code that was in the install but was not getting executed

FlexIDSCore #4811
Fixed problem with the installation of SQL CE, it will now check if the installed version is SP2 and if not will run the SQL CE SP2 install.

FlexIDSCore #4659
Added an FKB Installer and added code in the install for FlexIndex and IDShield to install.  Currently the Uninstall will not prompt to uninstall the FKB and they will have to be uninstalled separately.   The installation of the ComponentData has been removed from the regular install

William

[git-vault-id] Extract$/Engineering@16199/17343