﻿[FlexIDSCore: 4823] REPM import statements no longer require quotes around the file name.
[LegacyRCAndUtils:6166] (and probably some other SCRs)
All FAM task and supplier configuration dialogs are now formatted “<object name> settings”. (For most objects, this means “Configure:” was removed from the beginning and “settings” was added to the end.

[LegacyRCAndUtils:6167]
Fixed FAM task names that did not have proper casing.

[DotNetRCAndUtils:697]
File timestamp condition will now perform range conditions.

[DotNetRCAndUtils:705]
The same last minute tweak that caused [DotNetRCAndUtils:697] also prevented comparisons based on the timestamp of another file from working.

[DotNetRCAndUtils:700]
Removed “Accessed” option from file timestamp condition.

[DotNetRCAndUtils:704]
On the timestamp condition, I have changed “equal to” and “not equal to” to “exactly equal to” and “not exactly equal to” so that it is more clear how comparisons will be performed. Also, for the relative time option, I have added an example to make it clear what the timestamp is being compared to. This will appear when the relative time option is selected.

[DonNetRCAndUtils:708]
The “Send to FAM” context menu option now has an icon.

[DotNetRCAndUtils:709]
Name of context menu file supplier is now “Files from context menu”.

[DotNetRCAndUtils:710]
Context menu supplier will now treat the root path setting as a complete path and not allow the context menu to display on “X:\\Input2” when the filter is “X:\\Input”.

[DotNetRCAndUtils:724]
Context menu supplier now checks that the path root filter option is populated with a valid path.

[DotNetRCAndUtils:725]
ESIPCService will now be installed with permissions that allow any user to start and stop it. The context menu file supplier will now start the service if it is not already running so that users will no longer be required to do anything with the ESIPCService in order to use the context menu file supplier.

Also: The context menu file supplier now defaults to a .tif filter and has help bubbles to explain the menu option name and root folder filter.

[DotNetRCAndUtils:720]
The extract image area task no longer fails when not appending output while using the <AreaID> tag.

[DotNetRCAndUtils:727]
Fixed p
[git-vault-id] Extract$/Engineering@16209/17353