﻿[FlexIDSCore:4774] Fixed issue where regex rule when run as a modifier would un-necessarily search for matches beyond the first match.
[FlexIDSCore:4072]
While I was at it, added an option to the regex rule to indicate whether to find only the first match. (When used as a modifier, this option is always treated as checked).

Added an automated test case to test this new option.

Also, fixed a reference problem that was failing the build.

Reviewed by:
William (0/2)

[git-vault-id] Extract$/Engineering@16317/17461