﻿[FlexIDSCore:4645] It is not valid to use a relative path to specify an rsd file. I have updated the FindingRuleCondition  memory leak rsd file file to use the <RSDFileDir> tag.
Despite a note here that indicates <RSDFileDir> did not solve the issue, I'm going to go ahead and mark this as fixed because I think most of the cases reported here are due to this issue. If there is any remaining cases found, the underlying cause will be different and I think it would be deserving of a different SCR anyway...

Per discussion with Arvind, rules will no longer be able to saved with relative paths. I have updated the property page all rules whose configuration includes a path or support of a dynamic file to require use of a tag (unless an absolute filename is used).

Review pending…

[git-vault-id] Extract$/Engineering@16568/17715