﻿[LegacyRCAndUtils:6139] Saved reports are now saved to the correct location.
[FlexIDSCore:4858]
When F2 is pressed when currently zoomed to the current selection, it will return to the last zoom level for the current page that was not a zoom to selection (if there is one). This allows F2 to be used to toggle zoom in/out from the current selection.

Reviewed by:
William (0/0)

[git-vault-id] Extract$/Engineering@16618/17765