﻿[FlexIDSCore:4865] Fixed issue that prevents the rule profiler from distinguishing rule-objects if the entire ruleset was copied then pasted as a template for a new ruleset if both the original version and copy are called in the same profiling session. (Copying and pasting rule objects within or between rulesets was never a problem).
Now, the first time a ruleset is saved under as a new filename, all its rule object GUIDs will be updated so that they are unique. In other words, if you do nothing else but copy and paste an rsd file, the problem would still occur if, for some reason, both rulesets are called when executing rules. But as soon as the copied rule object is opened and saved (even if nothing else changed), new GUIDs will be generated for all its rule objects.

Reviewed by:
William (0/0)

[git-vault-id] Extract$/Engineering@16750/17900