﻿[LegacyRCAndUtils:6195] Due to database schema changes since 8.0, some file selection queries that were previously used in the FAMDBAdmin's SelectFiles custom query box have become substantially more complicated.
To limit the cases where custom queries need to be used in 9.0, the SelectFiles dialog now allows the existing conditions to be combined (either inclusively or exclusively). This will allow, for example, files to be selected based on their status in two different actions without needing to create a custom SQL query. Also, when custom SQL queries are necessary, in most cases they can be simpler than they would otherwise need to be because they can be combined with any of the other select file conditions.

[LegacyRCAndUtils:6192]
As part of this change, I have changed the query condition header from:
“SELECT FAMFile.FileName, FAMFile.ID FROM”
to
“SELECT FAMFile.FileName, FROM FAMFile”

This eliminates confusion regarding the ID column which is not output, but is used internally to sort. FAMFile is included because it is required to be referenced anyway. (Any query which had FAMFile as a join table should be able to be translated such that it is the primary table.)

Reviewed by:
William (2/1)
Follow-up review pending…

[git-vault-id] Extract$/Engineering@16836/17992