﻿[FlexIDSCore:4896, DataEntry:1071] In the IDShieldData and DataEntryData tables an “OverheadTime” column has been added.  The verification tasks now use it to record the amount of time passed since the "Duration" timer for the previous document was stopped unless the queue was empty after saving the last document (in which case the OverheadTime value will be zero for the next loaded document). Taken together, the Duration + OverheadTime columns should account for all of the time that the user was actively verifying.
The TotalDuration column has been removed since it can be calculated as part of the SQL query.

[FlexIDSCore:4895, DataEntry:1070]
When the processing tasks are configured to "Keep processing files as they are added to the queue", processing can now be stopped by closing the verification window or using the File|Stop processing menu item even when the queue is empty and no document is open.

Also, when the processing tasks are configured to "Keep processing files as they are added to the queue" and the slideshow is running, the slideshow will now stop immediately after the queue becomes empty rather than stopping the next time a document is loaded based upon the amount of time that has elapsed since the last document was closed.

Reviewed by:
William (3/9)

[git-vault-id] Extract$/Engineering@17024/18186