﻿[DotNetRCAndUtils:717] Added option in the FTP task to allow empty folders to be deleted if a deleted file was the last one in the directory.
[DotNetRCAndUtils:739]
Due to the fact that the after-download action on the FTP file supplier wasn’t atomic, to avoid any potential problems it has been removed.

[DotNetRCAndUtils:740]
Because the removed FTP supplier after-download action is now to be accomplished using the FTP task, a rename option has been added to the task needs to be able to replace what was previously done via the supplier.

Also:
- Updated the FTP event logging to be able to track directory events (for the delete empty folder option), and fixed at least once case where a failed FTP event wouldn’t be logged (when the local file couldn’t be found for an upload operation).
- Made the Retry class more generic by allowing any lambda expression to be used rather than requiring a pre-defined delegate.

Reviewed by:
William (2/1)

[git-vault-id] Extract$/Engineering@17114/18277