﻿[LegacyRCAndUtils:6154]
Added DB locks for a couple of calls at the start of processing where it appears there may have been a race condition that was leading to problems in the ActiveFAM table.

There are additional problems that don’t seem like they would relate directly back to problems in the ActiveFAM table, so I’m not going to mark this issue as fixed yet…. I think the thread scalability test should be repeated starting with build 114 and we’ll go from there.

Reviewed by:
William (1/0) (This entire change is basically William’s)

[git-vault-id] Extract$/Engineering@17181/18344