﻿[LegacyRCAndUtils:6120] Resolved potential database concurrency issue by modifying lock behavior such that a lock cannot be shared with additional threads in the same FileProcessingDB instance.
This change exposed a thread safety issue in the “Standby” code that, while apparently unlikely to occur without the above change, I think would be a good fix to have as part of the build used for Old Republic.

Therefore, I’m checking this in pre-review in order to get it into build 115.

Reviewed by:
William (2/0)

Second review pending…

[git-vault-id] Extract$/Engineering@17220/18384