﻿[FlexIDSCore:4913] Change Extract image area OH default to “Document context” rather than “Original image”
Also added an info tip bubble to warn that a uss file must be present for the “Original image” option.

[FlexIDSCore:4918, 4919]
Standardized all lists of multiple objects (rules objects, FAM conditions, etc) such that the name of the button to add a new item is always “Add…” (instead of “Insert…” or “Add”), and so that the new item is added after the currently selected item rather than before it.

Reviewed by:
William (0/1)

[git-vault-id] Extract$/Engineering@17465/18724