﻿[FlexIDSCore:4393] Fixed exception changing pages after rotating page in multi-page doc.
[DataEntry:1047]
Fixed weird swiping results when swiping a paragraph into a single-line field with the word highlight tool.

[DataEntry:1049]
When swiping into a single-line field with the word highlight tool, carriage returns are now preserved.

[DataEntry:1050]
Fixed exception generated using word highlight tool in a specific instance.

[DataEntry:1051]
Fixed issue where a swipe with the word highlight tool would result in a tooltip that displays upside down.

[DataEntry:1055]
Word highlight tool now uses the document OCR rather than attempting to re-OCR, except when in auto-fit mode (shift state). This fixes situations where words are well OCR’d on the document, but would not OCR correctly when selected with the word highlight tool.

[DataEntry:1068]
AutoOcr now works in DataEntry applications. A new tab is available in the Tools | Options… dialog that allows configuration of AutoOcr.

[DataEntry:1072]
Word highlight tool shift state (auto-fit mode) is once again available in data entry. This mode now has a different cursor to mirror the cursor for IDShield verification. Also, the auto-mode has been changed so that it always produces a hybrid string with the original bounds of the swipe, regardless of what OCRs (i.e., even if nothing OCRs). This provides an easy way to assign spatial information to a field regardless of whether any of its content will OCR.

Also, in a multi-DEP configuration the in-active DEPs no longer respond to image viewer events (which may have been dragging on performance in multi-DEP configurations).

Reviewed by:
William (0/0)

[git-vault-id] Extract$/Engineering@17477/18757