﻿[FlexIDSCore:4834] Fixed redaction resizing issue that could cause redaction zone heights to shrink when releasing the mouse at the end of a resize operation. NOTE: I added #4928 to regards to a separate situation where some very small redaction shrinking can be observed upon release when zoomed way in.
[FlexIDSCore:4709]
Keyboard shortcuts for open/close image are now available in ID Shield On-Demand.

[FlexIDSCore:4710]
The extension is now properly added when saving redacted output from ID Shield On-Demand.

[FlexIDSCore:4711]
The comments box in IDShield verification is now hidden if the UI is not connected to a file processing database.

[FlexIDSCore:4729]
Fixed problem with small redactions not appearing the same after saving and re-opening the image in verification.

Reviewed by:
William (1/0)

[git-vault-id] Extract$/Engineering@17502/18795