﻿[LegacyRCAndUtils:6217] Fixed redaction WSYWIG between sessions and resizing issues that resulted from converting RasterZones between the c++ and .Net size.
Redaction resizing testing note:

Redactions should now always appear exactly the same in all cases after saving and re-opening in redaction verification. Furthermore, some unnecessary bobble when resizing a redaction for the first time should now be corrected. For example, if the left or right side of a perfectly horizontal redaction is moved, it will no longer trigger the top and bottom sides to move.

However, I will point out that some bobble is still expected when releasing the mouse at the end of a resizing operation, particularly if there is any skew. This bobble will not always be "out"-- that is it may result in a slightly smaller zone. This is to prevent the frustration of all sides moving out in response to one side being adjusted. The side being moved should error on the side of moving out, though in some cases, even it may even appear to move very slightly if you are zoomed in enough.

Reviewed by:
William (0/0)

[git-vault-id] Extract$/Engineering@17519/18822