﻿[FlexIDSCore:4869] The exception being thrown is due to a COM call is failing (a com_error is being caught). It probably didn't get into the COM call or we would have exception trace information from within the call. I have added MLI codes to CAFEngineFileProcessor::raw_ProcessFile to attempt to figure out which COM call is failing.
[LegacyRCAndUtils:5945]
Added temporary logging to figure out why the IdleProcessKiller is not killing crashed/hung SSOCR2 instances. (Also added an SCR as a reminder to remove this temporary code.

[FlexIDSCore:4931, 4933]
New redactions are now added to the data pane grid in the order in which they appear in the document. This resolves tab order issues which could result in tab navigation getting caught in an infinite loop.

[FlexIDSCore:4932]
Read only items in the data grid view pane (Clues, by default) will now have a gray back ground when unselected and a gray text color when selected. The gray color is now a constant color instead of a theme color which will prevent it from being to light/dark depending on the OS and current theme.

[DotNetRCAndUtils:750]
If an FTP download fails, it will no longer overwrite a file that may have already existed in the destination location.

[DotNetRCAndUtils:751]
Path tags for “Local Working Folder” in the FTP file supplier are now evaluated for each file at the time it is downloaded.

[DotNetRCAndUtils:752]
Added a path tags button for the remote folder field in the FTP file supplier.

Reviewed by:
William (0/2)

[git-vault-id] Extract$/Engineering@17633/19043