﻿[LegacyRCAndUtils:6154] After applying my original "fix", in testing this afternoon, I did encounter another an error that had been logged in the original SCR regarding thread scalability issues so I have re-opened this SCR.
This exception is apparently due to the correct ActiveFAM record not being found in when updating the last ping time.  I am not sure how to truly solve this issue just yet as I have not figured out why the ActiveFAM record should be missing. But I do feel that a missing ActiveFAM entry should be a fatal error in regards to processing, which it wasn’t previously.

I have added indefinite retries for the pinging of the ActiveFAM table to allow for lost connections, and added code to allow for it to take out a lock to update the ActiveFAM table if necessary, but if the update still fails after taking out a lock on a valid connection, it will now result in processing being halted for the FAM.

[git-vault-id] Extract$/Engineering@17666/19076