﻿Reviewed by: William (2/0)
Fixed 2 issues in response to review. The biggest change is that the loss of its entry in the ActiveFAM table is no longer fatal to the FAM process; it will restore the entry in the ActiveFAM table and move on. This may result in some files getting processed by 2 different processes, but this situation should not be any different that with 8.0. Also note that there appear to be other circumstances that will cause a FAM process to fail after losing the DB connection for long enough.

[git-vault-id] Extract$/Engineering@17691/19101