﻿[LegacyRCAndUtils:6233] Between the endurance test and testing I have done in my configuration where I am running 50 processes X 2 threads plus a couple FAMs running separately, I have found that the primary database performance bottleneck is when instances need to lock the database. Queries to add and remove entries from the LockTable are at the top of the recent expensive queries list in terms of average duration. This is in addition to the obvious cost of holding up other processes that need the lock.
By default, the FAM’s stats thread needs to lock the database every 5 seconds. Both in the tests on my machine and in the endurance test, the call that seems to be responsible for the majority of the database locks is this one. In fact, in the testing on my machine, the FAMs I have running almost perpetually have a yellow database icon to indicate it is waiting on a lock. In the endurance test, this has been leading  to many “Timeout waiting for DB lock” exceptions displayed from the FAM stats thread.

Therefore, I have been experimenting today with having the updateActionStatisticsFromDelta triggered every 5 minutes instead of every 5 seconds, and rather than loadStats returning only what is in the ActionStatistics table, having it return an aggregation of the stats in both tables.

This seems to be working very well. Even with tens of thousands of rows in the delta table, the stats query is returning virtually instantaneously. With 2 FAMs running it shows up in the SQL mgmt studio’s expensive queries list, but with 0 in the average ms column (I’m sure there’s some cost, but not enough to register). When I run the query in SQL mgmt studio, the query returns instantly there as well.

In the testing I have been doing on my machine where I’m queuing with the 50 X 2 running and ~ a million files in the DB, I tried running with 10 stats FAMs (each updating once per second). I did notice in this case that the average ms cost according to SQL mgmt studio rose up to about 200 ms just before the updateActionStatisticsFromDelta call, but I did not observe any slowdown in queuing at all. (hovers between 30 to 40 files queued per second with or without the stats FAMs).

Further, since every call to load stats is getting the most up-to-date stats, rather that have
[git-vault-id] Extract$/Engineering@17846/19258