﻿ $/Engineering/ProductDevelopment/AttributeFinder/AFValueFinders/Code $/Engineering/ProductDevelopment/AttributeFinder/Build $/Engineering/ProductDevelopment/AttributeFinder/IndustrySpecific/Redaction/Installation/IDShield $/Engineering/ProductDevelopment/AttributeFinder/IndustrySpecific/Redaction/Installation/IDShield/Script Files $/Engineering/ProductDevelopment/AttributeFinder/Installation/UCLID FlexIndex SDK $/Engineering/ProductDevelopment/AttributeFinder/Installation/UCLID FlexIndex $/Engineering/ProductDevelopment/Common/InstallScripts $/Engineering/ProductDevelopment/DataEntry/LabDE/Installation/LabDE $/Engineering/ReusableComponents/COMComponents/ESImageCleanup/Code
Added following to vault
$/Engineering/ReusableComponents/APIs/ClearImage_7_0
$/Engineering/ReusableComponents/APIs/ClearImage_7_0/Bin
$/Engineering/ReusableComponents/APIs/ClearImage_7_0/Help
$/Engineering/ReusableComponents/APIs/ClearImage_7_0/Include
$/Engineering/ReusableComponents/APIs/ClearImage_7_0/Installer

FlexIDSCore #4400
Added support for ClearImage 7.0 and removed support for ClearImage 5.7

NOTE: On a developer machine you will need to unregister all of the the dll's in the RC/APIs/Inlite_5_7/bin folder and register the dlls in the RC/APIs/ClearImage_7_0/Bin folder.  You can use the following command in from dos prompt in the bin folder
to unregister:
     for /r %i in (*.dll) do regsvr32 %i /s /u

To register:
    for /r %i in (*.dll) do regsvr32 %i /s

Instead of registering the dlls for 7.0 you could run the Installer in the ClearImage_7_0\Installer folder

According to the Inlite salesman  you should be able to plug in the key with the license you want to uses an be able to run the operations for that license.  I had to actually use their license utility that is installed with the PDK to make it work on my machine but I had previously had the SDK and PDK installed on my machine and I also had a virtual machine that I setup to use the MICR key so this may have been why I had to use the license utility to select a license.

Reviewed by
Steve(0/0)

[git-vault-id] Extract$/Engineering@17916/19328