﻿[LegacyRCAndUtils:6242] Updated DB schema to include new table QueuedActionStatusChange. This table allows for the set file action file processor and redaction verification file processor to schedule the next processing state for currently processing files rather than failing.
For instance, if the Redact action is set to pending by the Verify action, but after navigating back to a previous document and re-saving the Redact action is already processing, the Verify task will schedule the action status of Redact to be set back to pending in the QueuedActionStatusChange and move on. When the file completes, fails, is skipped or is cancelled in the Redact task, rather than having its action be set as normal, the queued change will override and force it back into the pending state.

As with the FAST table, all  rows are maintained for problem investigation. Any action status change that was overridden by a queued status change will have a foreign key that points back to the corresponding QueuedActionStatusChange row and a FAST table comment will be added specifying the status that was overridden. For example: “Transition to C overridden”.

For any action status changes that do not arise from the file processing workflow, the QueuedActionStatusChange table will be ignored. This essentially means using the DBAdmin to set file action status for files that are actively processing.

The QueuedActionStatusChange has a Status column that will be set to “P” for queued changes that are pending, “C” for status changes that  have been completed, “O” for queued changes that have been overridden by a more recent QueuedActionStatusChange entry or “I” for queued changes that were ignored by a non-file-processing status change.

[LegacyRCAndUtils:6251]
When the GFTP gatekeeper query was introduced, it caused a bug that prevented stuck files from being auto-reverted if there were no pending files to be processed.

Also, fixed an apparent SQL syntax error in FileProcessingDB::RemoveFolder… which is apparently never called.

Review pending…

[git-vault-id] Extract$/Engineering@18022/19438