﻿Previous checkin reviewed by: William (1/2)
Checking in changes based on William’s review. The biggest change deals with the fact that the “fix” to gatekeeper query (below) had a marked performance impact on my 50X2 ThreadScalabilityTest. While I think the performance may still have been marginally acceptable given this is test is at the upper end of DB stressing we would plan for in production, per discussion with William a better resolution than the gatekeeper is to simply move the auto-revert code from GFTP to the ping thread since the auto-revert code was responsible for the lion’s share of GFTP’s performance hit.

The result is that even with the GFTP gatekeeper back out, DB performance is generally better in the 50x2 test than it was in builds 149-152.

Reviewed by:
William (0/1)

[git-vault-id] Extract$/Engineering@18037/19464