﻿[DataEntry:1073] Fixed issue where if the DataEntryApplication auto-zoom setting was set to “zoom out if necessary”, for the first field selected it would zoom all the way in to it.
[DataEntry:1074]
When a uss file was present for any Extract application that used the .Net ImageViewer, the uss file data was being loaded in an inefficient fashion. On documents with a lot of pages, this could cause a very noticeable delay loading the document.

[LegacyRCAndUtils:6222]
Removed temporary logging to track problems with the IdleProcessKiller; no problems with it were ever found.

[LegacyRCAndUtils:6223]
Upated copyright date in LogProcessStats.

Reviewed by:
William (1/1)

[git-vault-id] Extract$/Engineering@18094/19532