﻿[LegacyRCAndUtils:6248] Fixed a spot in COMUtils::readObjectFromFile which was supposed to have been retrying for files that couldn’t be accessed per LegacyRCAndUtils:5090, but wasn’t due to a small mistake. Also, I modified the LeadUtils::fillImageArea call so that it now makes 3 attempts for all failures, not just specific failures as it used to. There changes should address some occasional failures we are seeing in the endurance test.
Also, the query to get the count of files in FAMFile has been changed to a much more efficient query. This allows the FAMDBAdmin summary tab to be updated almost instantly on a multi-million record database, rather than taking minutes (this also means that SetFileAction status operations complete much faster since the refresh was a significant part of the time spent on the operation).

Reviewed by:
William (0/0)

[git-vault-id] Extract$/Engineering@18107/19545