﻿[DotNetRCAndUtils:742, 764], [LegacyRCAndUtils:6269]
In light of the problem Leadtools 17 has displaying some PDFs blank (and in the poor rasterization of some color PDFs), ImageFormatConverter now supports an “alternate method” of conversion via use of the new command line parameter “/am”. The “alternate method” is the Nuance engine. Likewise, the rasterize PDF task also supports an alternate method, but because Nuance performs better than Leadtools, the default method for the rasterize PDF task is Nuance, while Leadtools is the alternate.

Review pending…

[git-vault-id] Extract$/Engineering@18337/19788