﻿[DotNetRCAndUtils:766] The rasterize PDF task will no longer delete a pre-existing file matching the output file name if the rasterization fails.
[FlexIDSCore:4708]
In addition to prompting to save the VOA file (when appropriate) , when closing a dirty document in ID Shield On-Demand a second prompt is now also displayed that asks whether to output the redacted document as well.

[FlexIDSCore:4747]
In ID Shield On-demand, when there is a VOA file to save to and the user deletes a redaction, it will no longer throw an exception on the second save attempt if the first save was cancelled before selecting the redacted output location.

Reviewed by:
William (0/0)

[git-vault-id] Extract$/Engineering@18373/19825