﻿[LegacyRCAndUtils:6174] Updated packages.dat file so that in includes 2 versions of the IDShield Client license: 1 with OCROnClient, and 1 without.
[DotNetRCAndUtils:758]
Added new time span options in the report viewer. Also, modified the Refresh functionality so that it re-interprets the selected parameters. For example, if a report is run for the “Last 10 minutes” at 12:00, it will be fore 11:50 to 11:51. If it is refreshed at 12:01, it will be for the range 11:51 to 12:01.

Reviewed by:
William (0/0)

[git-vault-id] Extract$/Engineering@18402/19854