﻿[FlexIDSCore:4901] Fixed issue where “Create redacted image” task would add the IDAndRevision metadata attribute to metadata attributes.
[FlexIDSCore:6184, 6206]
I wasn’t able to reproduce or see a possible cause for the exceptions logged from CreateSearchablePDF in an endurance test;  however I made a couple tweaks (including adding a retry) that I hope might possibly prevent the issue (though I’m not too hopeful) In any case, these exceptions are not indicative of a critical problem; at worst some memory hasn’t been freed and a temp file was not deleted.

Arvind asked that I review the project settings for all c++ projects to ensure they are not set to be optimized as it may relate to the FAM crash issue. I confirmed that optimizations were turned off for all probjects, however I did find and correct a couple of projects that were not using the c++ exceptions setting that we should be using:  “Yes with SHE Exceptions (/EHa)”.

Reviewed by:
William (0/0)

[git-vault-id] Extract$/Engineering@18421/19873