﻿[FlexIDSCore:6282] Fixed a problem that would cause an out of memory error when attempted to set file action status on a subset of the database that is a least several million records in size. Included is another change that not only prevents large numbers of skipped files from drastically reducing the speed of the altered algorithm but fixes a situation in which stats could be corrupted by setting the file action status from skipped to another status while some/all of those skipped files are processing in a FAM that is processing skipped files.
[Note the stats issue with skipped files was due in part to the fact that when setting file action status from skipped to another status, the query doesn’t actually work of the file’s real current status, but whether that file is in the skipped file table. This doesn’t seem right to me and may be a bug in its own right.]

[FlexIDSCore:6284]
The query the FAMDBAdmin was using to retrieve the number of files in the DB very quickly was turned out to require admin rights in the database. I am still trying the fast method first, but if it fails with a permissions error, an app trace is logged and the standard query is used.

[LegacyRCAndUtils:6285]
Fixed a problem where an FPS file configured to processed skipped files by any user would fail if processing had not yet been run since loading.

[git-vault-id] Extract$/Engineering@18545/19997