﻿[FlexIDSCore:5000] Removed an exception that would be logged in association with the word highlight tool when pages were rapidly changed for more than a second.
The only bad effect of the situation for which the exception was getting logged is that a handle is not being deterministically disposed of.  However, garbage collection should still take care of it. The only way to truly prevent the situation would be to wait indefinitely, but this would risk the worse possibility of the word highlight tool becoming unusable.

Reviewed by:
William (0/1)

[git-vault-id] Extract$/Engineering@18616/20068