﻿[FlexIDSCore:5011, 5026] Fixed zoom and fit mode behavior so that it works correctly (with or without tile navigation).
[FlexIDSCore:5015]
If the thumbnail window is used in auto-hide mode, the thumbnails should now be available at all times.

Fixed a most likely innocuous situation with occurred with a standby thread in on one of the endurance tests that led to an exception being logged. I do not think this would have related to any crash/corruption issues, but I fixed it to be on the safe side.

Review pending…

[git-vault-id] Extract$/Engineering@18693/20152