﻿[FlexIDSCore:5002] Fixed but in RSD scorer where score could be incorrect if one attribute name was a subset of another variable name.
NOTE TO RULEWRITES:
If there are any real-world implementations of the RSD data scorer where you have used multiple attribute names and one of the attribute names is part of another attribute name (e.g. “test” and “test1”), those need opened back up, and edited (add a space to the end and save it).

Reviewed by:
William (0/0)

[git-vault-id] Extract$/Engineering@18736/20197