﻿[FlexIDSCore:5037, 3906] Another attempt at fixing a tight message loop (leading to a freeze) that still wasn’t completely fixed in my last attempt.
[FlexIDSCore:5057, 5062]
Fixed bug in ID Shield On Demand that threw an exception when saving.

[FlexIDSCore:5058]
Add check to prevent slideshow from starting during a tracking event.

[FlexIDSCore:5063]
Next document button and shortcut key is not always disabled when on the current document.

[FlexIDSCore:5066]
You can no longer use shortcut keys to navigate to a page that doesn’t exist in the page summary grid.

[FlexIDSCore:5068]
Fixed exception message displayed by find and redact feature when OCR is not available.

Review pending…

[git-vault-id] Extract$/Engineering@18836/20305