﻿[DataEntry:1117] Added redo feature to go along with undo.
I have added a SCR to improve upon the handling of when the redo stack gets cleared. Currently it gets cleared for any data change which includes viewed/unviewed status. This means simply selecting another field can clear the redo stack. I think users could find this frustrating, but it was not trivial to address, so I have left it as an SCR for now.

Reviewed by:
William (0/1)

[git-vault-id] Extract$/Engineering@19267/20836