﻿RC.Net\DataEntry RC.Net\UtilityApplications\SQLCompactUtils\SQLCDBEditor
[DotNetRCAndUtils:817]
Fixed problem that prevented erred rows from being displayed when attempting to save or clicking one the invalid data label if the rows in the table have been re-sorted since the data became invalid. Also, when updates from other tables are applied to tables which have had sorting changed, the sorting will no longer be reset by the update.

[DotNetRCAndUtils:818]
User is now correctly prompted to save dirty queries even when then don’t yet exist on disk.

[DotNetRCAndUtils:819]
Solved the potential for concurrency errors when making updates to data in one table that is used as a foreign key in another table that contains error(s).
When a table contains errors, changes made to keys in other tables will still be applied to the table with invalid data despite the errors. If the change happens to resolve the error(s), the errors will not be cleared without needing to modify the erred row(s).

Also:
- So that the updated utility can be dropped into an existing installation, temporarily copied some database utility methods into the SQLCDBEditor project so that it is not dependent on post-9.0 changes.
- Solved a couple cases where the selection in the table or query list box was not consistent with the currently active query or table tab. (e.g., the list box selection would remain when there was no active table or query)
- Changed some icons per discussion with Arvind.
- Fixed a control alignment issue.

[git-vault-id] Extract$/Engineering@19805/21402