﻿HNL DEP changes: - Fixed query to allow the flag value to be auto-updated even after it already has a value. - Fixed query so that the order name will be auto-populated when and order code is entered in the order code column (this query was never changed to reflect the fact that HNL’s DEP is displaying a different code column that the standard LabDE Panel). - The value for an “UNDER DIRECTION OF” test is now required to be one of the physicians in the physician table. - Changed the value of the mandatory test validation query for orders so that it is compatible with the post 9.0 query work, then added an override of the query in the HNL config files so that I can use the same version of the DEP HNL is using on my dev machine.
[DataEntry:1122]
Changed query in the StandardLabDE panel that had been tailored to an AH specific requirement—that only physicians with a 4 or 5 digit code be auto-populated in the physician tables. The standard LabDE panel will now auto-populate physician names regardless of the number of digits in the code.

I have updated the network folder to with these changes:
I:\Common\Engineering\ProductDevelopment\AttributeFinder\Customer Specific Files\Health Network Labs\DeployedFPSAndOtherFiles\Solution

[git-vault-id] Extract$/Engineering@19841/21438